requirements.txt
beautifulsoup4==4.13.4
curl_cffi==0.12.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.4